requirements.txt
google-generativeai
st-gsheets-connection
pandas